+++ mkchromecast/__init__.py
@@ -8,7 +8,7 @@ import mkchromecast.colors as colors
 from mkchromecast.terminate import *
 from mkchromecast.version import __version__
 import os.path, sys, platform
-import pickle
+import pickle, subprocess
 from argparse import RawTextHelpFormatter
 
 parser = argparse.ArgumentParser(description='Cast Mac OS X and Linux audio to your Google Cast devices.', formatter_class=RawTextHelpFormatter)
@@ -42,7 +42,7 @@ Possible codecs:
     - wav  [HQ]     Waveform Audio File Format
     - flac [HQ]     Free Lossless Audio Codec
 
-This option only works for the ffmpeg backend.
+This option only works for the ffmpeg and avconv backends.
 
 ''')
 parser.add_argument('--config', action="store_true", help='Use this option to connect from configuration file')
@@ -54,13 +54,23 @@ Set the backend for all encoders.
 Possible backends:
     - node (default)
     - ffmpeg
+    - avconv
 
 Example:
     python mkchromecast.py --encoder-backend ffmpeg
 
 ''')
 parser.add_argument('-n', '--name', action="store_true", help='Use this option if you know the name of the Google Cast you want to connect')
-parser.add_argument('-r', '--reset', action="store_true", help='When the application fails, and you have no audio in your laptop, use this option to reset')
+parser.add_argument('--notifications', action="store_true", help='''
+Use this flag to enable the notifications.
+''')
+parser.add_argument('-r', '--reset', action="store_true", help='''
+When the application fails, and you have no audio in your computer, use this
+option to reset the computer's audio
+''')
+parser.add_argument('--reboot', action="store_true", help='''
+Reboot the Google Cast device
+''')
 parser.add_argument('-s', '--select-cc', action="store_true", help='If you have more than one Google Cast device use this option')
 parser.add_argument('--sample-rate', type=int, default='44100', help=
 '''
@@ -70,7 +80,7 @@ Audio MIDI Setup in the "Soundflower (2ch)" audio device. You need to change
 the "Format" in both input/output from 44100Hz to maximum 96000Hz. I think that
 more than 48000Hz is not necessary, but this is up to the users' preferences.
 
-Note that resampling to higher sample rates is not a good idea. It was indeed
+Note that re-sampling to higher sample rates is not a good idea. It was indeed
 an issue in the chromecast audio. See: https://goo.gl/yNVODZ.
 
 Example:
@@ -96,8 +106,21 @@ For more information see: http://wiki.audacityteam.org/wiki/Sample_Rates.
 ''')
 parser.add_argument('-t', '--tray', action="store_true", help=
 '''
-This option let you launch mkchromecast as a systray menu (still experimental)
+This option let you launch mkchromecast as a systray menu (beta)
 ''')
+parser.add_argument('--update', action="store_true", help="""
+Update mkchromecast git repository.
+
+Example:
+    python mkchromecast.py --update
+
+This will execute for you:
+
+    git pull --all
+    git fetch -p
+
+"""
+)
 parser.add_argument('-v', '--version', action="store_true", help='Show the version')
 parser.add_argument('--volume', action="store_true", default=False, help=
 '''
@@ -124,6 +147,20 @@ Guess the platform
 platform = platform.system()
 
 """
+Assigment of args to variables
+"""
+tray = args.tray
+if tray == True:    # This should fix the problem with the Mac app
+    select_cc = True
+else:
+    select_cc = args.select_cc
+debug = args.debug
+if args.notifications == True:
+    notifications = 'enabled'
+else:
+    notifications = 'disabled'
+
+"""
 Reset
 """
 if args.reset == True:
@@ -140,6 +177,20 @@ if args.config == True or args.discover == True or args.name == True:
     sys.exit(0)
 
 """
+Reboot
+"""
+if args.reboot == True:
+    print (colors.error('This option is not implemented yet.'))
+    sys.exit(0)
+
+"""
+Not yet implemented
+"""
+if args.config == True or args.discover == True or args.name == True:
+    print (colors.error('This option is not implemented yet.'))
+    sys.exit(0)
+
+"""
 Version
 """
 if args.version is True:
@@ -147,20 +198,44 @@ if args.version is True:
     sys.exit(0)
 
 """
+Update
+"""
+if args.update is True:
+
+    print (colors.warning('Updating mkchromecast'))
+    print (colors.important('git pull --all'))
+    pull = subprocess.Popen(['git', 'pull', '--all'], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    print (pull.stdout.read().decode('utf-8').strip())
+    print (colors.important('git fetch -p'))
+    prune = subprocess.Popen(['git', 'fetch', '-p'], stdout=subprocess.PIPE, stderr=subprocess.PIPE  )
+    print (prune.stdout.read().decode('utf-8').strip())
+    sys.exit(0)
+
+"""
 Check that encoders exist in the list
 """
-backends = ['node', 'ffmpeg']
+backends = ['node', 'ffmpeg', 'avconv']
 
-if args.encoder_backend in backends:
-    backend = args.encoder_backend
-    if platform == 'Linux' and backend == 'node':
-        args.encoder_backend = 'ffmpeg'
-        backend = args.encoder_backend
-else:
+if args.encoder_backend not in backends:
     print (colors.error('Supported backends are: '))
     for backend in backends:
         print ('-',backend)
     sys.exit(0)
+else:
+    if platform == 'Darwin':
+        backends.remove('avconv')
+    else:
+        backends.remove('node')
+
+    if args.encoder_backend in backends:
+        backend = args.encoder_backend
+    elif args.encoder_backend not in backends:
+        if platform == 'Linux':
+            args.encoder_backend = 'ffmpeg'
+            backend = args.encoder_backend
+        elif platform == 'Darwin':
+            args.encoder_backend = 'node'
+            backend = args.encoder_backend
 
 """
 Debug
@@ -239,15 +314,15 @@ if args.youtube != None:
 This is to write a PID file
 """
 def writePidFile():
-    if os.path.exists('/tmp/mkcrhomecast.pid') == True:     #This is to verify that pickle tmp file exists
-       os.remove('/tmp/mkcrhomecast.pid')
+    if os.path.exists('/tmp/mkchromecast.pid') == True:     #This is to verify that pickle tmp file exists
+       os.remove('/tmp/mkchromecast.pid')
     pid = str(os.getpid())
-    f = open('/tmp/mkcrhomecast.pid', 'wb')
+    f = open('/tmp/mkchromecast.pid', 'wb')
     pickle.dump(pid, f)
     f.close()
     return
 
 def checkmktmp():
-    if os.path.exists('/tmp/mkcrhomecast.tmp') == True:     #This is to verify that pickle tmp file exists
-       os.remove('/tmp/mkcrhomecast.tmp')
+    if os.path.exists('/tmp/mkchromecast.tmp') == True:     #This is to verify that pickle tmp file exists
+       os.remove('/tmp/mkchromecast.tmp')
     return