+++ mkchromecast/cast.py
@@ -6,19 +6,33 @@ from __future__ import print_function
 from mkchromecast.__init__ import *
 from mkchromecast.audiodevices import *
 import mkchromecast.colors as colors
+from mkchromecast.config import *
+from mkchromecast.preferences import ConfigSectionMap
 from mkchromecast.terminate import *
 import time
 import pychromecast
+from pychromecast.dial import reboot
 import socket
 import os.path
 import pickle
 import subprocess
 import mkchromecast.colors as colors
+"""
+Configparser is imported differently in Python3
+"""
+try:
+    import ConfigParser
+except ImportError:
+    import configparser as ConfigParser # This is for Python3
 
 class casting(object):
     def __init__(self): ## __init__ to call the self.ip
         import mkchromecast.__init__        # This is to verify against some needed variables
         self.platform = mkchromecast.__init__.platform
+        self.tray = mkchromecast.__init__.tray
+        self.select_cc = mkchromecast.__init__.select_cc
+
+        self.backend = mkchromecast.__init__.backend
         if self.platform == 'Linux':
             hostname = subprocess.Popen(['hostname', '-I'], stdout=subprocess.PIPE)
             self.ip = hostname.stdout.read().decode('utf-8').strip()
@@ -26,8 +40,8 @@ class casting(object):
             self.ip = socket.gethostbyname(socket.gethostname())
 
     def initialize_cast(self):
-        from pychromecast import socket_client
         import mkchromecast.__init__        # This is to verify against some needed variables
+        from pychromecast import socket_client
         self.cclist = list(pychromecast.get_chromecasts_as_dict().keys())
 
         try:
@@ -35,7 +49,7 @@ class casting(object):
         except AttributeError:
             self.youtubeurl = None
 
-        if len(self.cclist) != 0 and args.select_cc == False:
+        if len(self.cclist) != 0 and self.select_cc == False:
             print (' ')
             print (colors.important('List of Google Cast devices available in your network:'))
             print (colors.important('------------------------------------------------------'))
@@ -51,9 +65,9 @@ class casting(object):
             print (colors.success(self.castto))
             print (' ')
 
-        elif len(self.cclist) != 0 and args.select_cc == True:
-            if os.path.exists('/tmp/mkcrhomecast.tmp') == False:
-                self.tf = open('/tmp/mkcrhomecast.tmp', 'wb')
+        elif len(self.cclist) != 0 and self.select_cc == True:
+            if os.path.exists('/tmp/mkchromecast.tmp') == False:
+                self.tf = open('/tmp/mkchromecast.tmp', 'wb')
                 print (' ')
                 print (colors.important('List of Google Cast devices available in your network:'))
                 print (colors.important('------------------------------------------------------'))
@@ -69,21 +83,21 @@ class casting(object):
                 #print (availablecc)
 
             else:
-                self.tf = open('/tmp/mkcrhomecast.tmp', 'rb')
+                self.tf = open('/tmp/mkchromecast.tmp', 'rb')
                 self.index=pickle.load(self.tf)
                 self.castto = self.cclist[int(self.index)]
                 print (' ')
                 print (colors.options('Casting to:')+' '+colors.success(self.castto))
                 print (' ')
 
-        elif len(self.cclist) != 0 and args.select_cc == False and args.tray == True :
-                self.tf = open('/tmp/mkcrhomecast.tmp', 'rb')
+        elif len(self.cclist) != 0 and self.select_cc == False and self.tray == True:
+                self.tf = open('/tmp/mkchromecast.tmp', 'rb')
                 self.index=pickle.load(self.tf)
                 self.castto = self.cclist[int(self.index)]
                 print (' ')
                 print (colors.options('Casting to:')+' '+colors.success(self.castto))
 
-        elif len(self.cclist) == 0 and args.tray == False:
+        elif len(self.cclist) == 0 and self.tray == False:
             print (colors.error('No devices found!'))
             if self.platform == 'Linux':
                 import mkchromecast.pulseaudio
@@ -94,8 +108,8 @@ class casting(object):
             terminate()
             exit()
 
-        elif len(self.cclist) == 0 and args.tray == True:
-            print (colors.error('No devices found!'))
+        elif len(self.cclist) == 0 and self.tray == True:
+            print (colors.error(':::Tray::: No devices found!'))
             self.availablecc = []
 
     def sel_cc(self):
@@ -127,7 +141,11 @@ class casting(object):
 
     def play_cast(self):
         localip = self.ip
+        #if platform == 'Darwin':
+        #   self.host = socket.gethostbyname(self.castto+'.local')
+        self.host = self.cast.host
 
+        print (colors.options('The IP of ')+colors.success(self.castto)+colors.options(' is:')+' '+self.host)
         print (colors.options('Your local IP is:')+' '+localip)
 
         if self.youtubeurl != None:
@@ -149,7 +167,20 @@ class casting(object):
             yt.play_video(video)
         else:
             ncast = self.cast
-            if args.encoder_backend == 'ffmpeg':
+
+            if self.tray == True:
+                config = ConfigParser.RawConfigParser()
+                configurations = config_manager()    # Class from mkchromecast.config
+                configf = configurations.configf
+
+                if os.path.exists(configf) and self.tray == True:
+                    print (tray)
+                    print(colors.warning('Configuration file exist'))
+                    print(colors.warning('Using defaults set there'))
+                    config.read(configf)
+                    self.backend = ConfigSectionMap("settings")['backend']
+
+            if self.backend == 'ffmpeg' or self.backend == 'avconv':
                 import mkchromecast.ffmpeg
                 mtype = mkchromecast.ffmpeg.mtype
                 print (' ')
@@ -160,7 +191,7 @@ class casting(object):
                 print (colors.options('The media type string used is:')+' '+  'audio/mpeg')
                 ncast.play_media('http://'+localip+':3000/stream.mp3', 'audio/mpeg')
             print (' ')
-            print (colors.important('Cast media cotroller status'))
+            print (colors.important('Cast media controller status'))
             print (' ')
             print (ncast.status)
             print (' ')
@@ -188,3 +219,11 @@ class casting(object):
         ncast = self.cast
         volume = round(ncast.status.volume_level, 1)
         return ncast.set_volume(volume - 0.1)
+
+    def reboot(self):
+        if self.platform == 'Darwin':
+            self.host = socket.gethostbyname(self.castto+'.local')
+            reboot(self.host)
+        else:
+            print (colors.error('This method is not supported in Linux yet.'))
+